D-Class_back/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9
django-extensions>=3.2.3

# CORS & Security
django-cors-headers>=4.3.1

# File Storage (AWS S3)
django-storages>=1.14.2
boto3>=1.34.0

# Filtering & Search
django-filter>=23.5

# Celery (Async Tasks)
celery>=5.3.4
redis>=5.0.1

# Environment Variables
python-decouple>=3.8

# Image Processing
Pillow>=10.2.0

# Date/Time
python-dateutil>=2.8.2

# Utilities
pytz>=2024.1

